revert: chore: reexport type and utils to shared (575008d5)
