+++ packages/maskbook/src/database/Persona/Persona.db.ts
@@ -1,5 +1,6 @@
 /// <reference path="../global.d.ts" />
 import Fuse from 'fuse.js'
+import { ProfileIdentifier, PersonaIdentifier, Identifier, ECKeyIdentifier } from '../type'
 import { DBSchema, openDB } from 'idb/with-async-ittr-cjs'
 import { IdentifierMap } from '../IdentifierMap'
 import { PrototypeLess, restorePrototype } from '../../utils/type'
@@ -7,18 +8,11 @@ import { MaskMessage } from '../../utils/messages'
 import { IDBPSafeTransaction, createTransaction, createDBAccessWithAsyncUpgrade } from '../helpers/openDB'
 import { queryProfile } from './helpers'
 import { assertPersonaDBConsistency } from './consistency'
-import {
-    ProfileRecord,
-    PersonaRecord,
-    LinkedProfileDetails,
+import type {
     AESJsonWebKey,
-    ProfileIdentifier,
-    PersonaIdentifier,
-    Identifier,
-    ECKeyIdentifier,
-} from '@dimensiondev/maskbook-shared'
-
-export type { ProfileRecord, PersonaRecord, LinkedProfileDetails }
+    EC_Public_JsonWebKey,
+    EC_Private_JsonWebKey,
+} from '../../modules/CryptoAlgorithm/interfaces/utils'
 import { CryptoKeyToJsonWebKey } from '../../utils/type-transform/CryptoKey-JsonWebKey'
 /**
  * Database structure:
@@ -486,6 +480,41 @@ export async function deleteProfileDB(id: ProfileIdentifier, t: ProfileTransacti
 //#endregion
 
 //#region Type
+export interface ProfileRecord {
+    identifier: ProfileIdentifier
+    nickname?: string
+    localKey?: AESJsonWebKey
+    linkedPersona?: PersonaIdentifier
+    createdAt: Date
+    updatedAt: Date
+}
+
+export interface LinkedProfileDetails {
+    connectionConfirmState: 'confirmed' | 'pending' | 'denied'
+}
+
+export interface PersonaRecord {
+    identifier: PersonaIdentifier
+    /**
+     * If this key is generated by the mnemonic word, this field should be set.
+     */
+    mnemonic?: {
+        words: string
+        parameter: { path: string; withPassword: boolean }
+    }
+    publicKey: EC_Public_JsonWebKey
+    privateKey?: EC_Private_JsonWebKey
+    localKey?: AESJsonWebKey
+    nickname?: string
+    linkedProfiles: IdentifierMap<ProfileIdentifier, LinkedProfileDetails>
+    createdAt: Date
+    updatedAt: Date
+    /**
+     * create a dummy persona which should hide to the user until
+     * connected at least one SNS identity
+     */
+    uninitialized?: boolean
+}
 type ProfileRecordDB = Omit<ProfileRecord, 'identifier' | 'hasPrivateKey' | 'linkedPersona'> & {
     identifier: string
     network: string
