+++ packages/maskbook/src/database/Persona/types.ts
@@ -1,5 +1,18 @@
-import type { PersonaRecord } from '@dimensiondev/maskbook-shared'
-export type { Profile, Persona } from '@dimensiondev/maskbook-shared'
+import type { ProfileRecord, PersonaRecord, LinkedProfileDetails } from './Persona.db'
+import type { ReadonlyIdentifierMap } from '../IdentifierMap'
+import type { ProfileIdentifier } from '../type'
+
+type TypedOmit<T, Q extends keyof T> = Omit<T, Q>
+export interface Profile extends Readonly<TypedOmit<ProfileRecord, 'localKey' | 'linkedPersona'>> {
+    readonly linkedPersona?: Persona
+    readonly avatar?: string
+}
+
+export interface Persona extends TypedOmit<PersonaRecord, 'localKey' | 'publicKey' | 'privateKey' | 'linkedProfiles'> {
+    readonly linkedProfiles: ReadonlyIdentifierMap<ProfileIdentifier, LinkedProfileDetails>
+    readonly hasPrivateKey: boolean
+    readonly fingerprint: string
+}
 
 export interface PersonaWithPrivateKey
     extends Omit<PersonaRecord, 'privateKey'>,
