--- packages/shared/src/database/index.ts
@@ -1,2 +1,0 @@
-export * from './Persona/types'
-export * from './Persona/Persona.db'
