+++ packages/shared/src/index.ts
@@ -10,6 +10,5 @@ export * from './ShadowRoot'
 export * from './TypedMessage'
 export * from './utils'
 export * from './hooks'
-export * from './database'
 export * from './crypto/JWKType'
 export * from './wallet'
